+++ src/components/product/ModalReview/index.tsx
@@ -7,24 +7,76 @@ import { FieldValues, FormProvider, useForm } from "react-hook-form";
 import ERROR_MESSAGE from "../../../constant/ERROR_MESSAGE";
 import FormTextareaInput from "../../common/input/FormTextareaInput";
 import FormMultiImageInput from "../../common/input/FormMultiImageInput";
+import { useMutation, useQueryClient } from "@tanstack/react-query";
+import { postReview } from "@/src/apis/review";
+import { QUERY_KEY } from "@/src/routes";
+import { PostReviewRequestType } from "@/src/apis/review/schema";
+import { postImage } from "@/src/apis/image";
 
 interface ModalReviewProps {
   productId: number;
   name: string;
   category?: CategoryType | undefined;
-  callback?: (data: FieldValues) => Promise<void>;
+  // callback?: (data: FieldValues) => Promise<void>;
   onClose: () => void;
 }
 
-function ModalReview({ name, category, onClose, callback }: ModalReviewProps) {
+function ModalReview({ productId, name, category, onClose }: ModalReviewProps) {
   const [score, setScore] = useState(5);
   const [reviewImages, SetReviewImages] = useState<string[]>([]);
 
   const methods = useForm();
+  const queryClient = useQueryClient();
+
+  // 리뷰 생성 요청
+  const postReviewMutation = useMutation({
+    mutationFn: (newReview) => postReview(newReview),
+    onSuccess: () => {
+      queryClient.invalidateQueries({ queryKey: [QUERY_KEY.REVIEWS] });
+    },
+  });
+
+  // 리뷰 생성 버튼 클릭시 발생 이벤트
+  const postReviewCallback = async (data: FieldValues) => {
+    const formData: PostReviewRequestType = {
+      productId: productId,
+      images: [],
+      content: data.content,
+      rating: data.rating,
+    };
+
+    // data 이미지 파일이 있을 경우 파일들을 업로드하고 새로운 URL 받아서 formData.images에 추가
+    const getImageUrlPromises = [];
+
+    if (data.images !== undefined) {
+      for (const file of data.images) {
+        let newImageUrl = await postImage(file);
+        getImageUrlPromises.push(newImageUrl);
+      }
+
+      try {
+        const imageUrlResponse = await Promise.all(getImageUrlPromises);
+        formData.images = imageUrlResponse.map((response) => response.url);
+      } catch (erroe) {
+        console.log("이미지 업로드를 실패했어요.");
+      }
+    }
+    postReviewMutation.mutate(formData, {
+      onSuccess: () => {
+        // console.log("리뷰가 성공적으로 업로드 되었습니다!");
+      },
+    });
+  };
 
   return (
     <FormProvider {...methods}>
-      <Modal title={name} modalType="review" category={category} onClose={onClose} callback={callback} isFormData>
+      <Modal
+        title={name}
+        modalType="review"
+        category={category}
+        onClose={onClose}
+        callback={postReviewCallback}
+        isFormData>
         <S.Container>
           <S.Rating>
             별점
